# src/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BadBeatMods API",
    "description": "This isn't really fully complete, but its better than absolutely nothing.\n\nThis API documentation is automatically generated and therefor may not be 100% accurate and may be missing a few fields.",
    "version": "0.0.1"
  },
  "host": "bbm.saera.gay",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json",
    "multipart/form-data",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/versions": {
      "get": {
        "tags": [
          "Versions"
        ],
        "description": "",
        "parameters": [
          {
            "name": "gameName",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Versions"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "version": {
                  "example": "any"
                },
                "gameName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/versions/default": {
      "get": {
        "tags": [
          "Versions"
        ],
        "description": "",
        "parameters": [
          {
            "name": "gameName",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Versions"
        ],
        "description": "",
        "parameters": [
          {
            "name": "gameVersionId",
            "description": "The ID of the version to set as default",
            "type": "number",
            "in": "query"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "gameVersionId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user information.",
        "description": "Get user information.",
        "responses": {
          "200": {
            "description": "Returns user information."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user information.",
        "description": "Get user information.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "User ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user information."
          },
          "400": {
            "description": "Invalid parameters."
          },
          "404": {
            "description": "User not found."
          }
        }
      }
    },
    "/api/user/{id}/mods": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user information.",
        "description": "Get user information.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "User ID."
          },
          {
            "name": "status",
            "description": "Status of the mod.",
            "type": "string",
            "in": "query"
          },
          {
            "name": "platform",
            "description": "Platform of the mod.",
            "type": "string",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns mods."
          },
          "400": {
            "description": "Invalid parameters."
          },
          "404": {
            "description": "User not found."
          }
        }
      }
    },
    "/api/user/{id}/": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Get user information.",
        "description": "Get user information.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "displayName": {
                  "example": "any"
                },
                "sponsorUrl": {
                  "example": "any"
                },
                "bio": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user information."
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    }
  }
}